Refactor: Extract common page assertion methods into base class
Since most of the expectations are shared for the boards index and
overview pages, I extracted these shared page expectations into a base
class that these page classes can inherit from and extend with their own
custom expectations/override base implementations.
