+++ modules/boards/spec/features/support/board_list_page.rb
@@ -27,17 +27,54 @@
 #++
 
 require 'support/pages/page'
-require_relative 'board_list_page'
 
 module Pages
-  class BoardOverview < BoardListPage
+  class BoardListPage < Page
     def visit!
-      navigate_to_modules_menu_item("Boards")
+      raise 'Define how to visit me'
     end
 
-    def expect_global_menu_item_selected
-      within '#main-menu' do
-        expect(page).to have_selector('.selected', text: 'Boards')
+    def expect_create_button
+      within '.toolbar-items' do
+        expect(page).to have_link 'Board'
+      end
+    end
+
+    def expect_no_create_button
+      within '.toolbar-items' do
+        expect(page).not_to have_link 'Board'
+      end
+    end
+
+    def expect_boards_listed(*boards)
+      within '#content-wrapper' do
+        boards.each do |board|
+          expect(page).to have_selector("td.name", text: board.name)
+        end
+      end
+    end
+
+    def expect_boards_not_listed(*boards)
+      within '#content-wrapper' do
+        boards.each do |board|
+          expect(page).not_to have_selector("td.title", text: board.name)
+        end
+      end
+    end
+
+    def expect_no_boards_listed
+      within '#content-wrapper' do
+        expect(page).to have_content I18n.t(:no_results_title_text)
+      end
+    end
+
+    def expect_to_be_on_page(number)
+      expect(page).to have_selector('.op-pagination--item_current', text: number)
+    end
+
+    def to_page(number)
+      within '.op-pagination--pages' do
+        click_link number.to_s
       end
     end
   end
